% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{accuracy_score}
\alias{accuracy_score}
\title{Accuracy Score}
\usage{
accuracy_score(y_true, y_pred)
}
\arguments{
\item{y_true}{true labels}

\item{y_pred}{predicted labels}
}
\value{
numeric accuracy
}
\description{
Computes classification accuracy.
}
\details{
Provides functionality for accuracy_score operations.
}
\examples{
  y_true <- c(1,0,1,1)
  y_pred <- c(1,0,0,1)
  accuracy_score(y_true, y_pred)

}
\seealso{
\code{\link{VectorForgeML-package}}
}
