% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_transformer.R
\docType{class}
\name{ColumnTransformer-class}
\alias{ColumnTransformer-class}
\alias{ColumnTransformer}
\title{Column Transformer}
\value{
ColumnTransformer object
}
\description{
Applies transformations to specific columns.
}
\details{
Provides functionality for ColumnTransformer operations.
}

\examples{
  model <- ColumnTransformer$new(num_cols="A", cat_cols="B")

}
\seealso{
\code{\link{VectorForgeML-package}}
}
