% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{confusion_stats}
\alias{confusion_stats}
\title{Confusion Matrix Statistics}
\usage{
confusion_stats(cm)
}
\arguments{
\item{cm}{confusion matrix}
}
\value{
list
}
\description{
Calculates accuracy, precision, recall, F1 from confusion matrix.
}
\details{
Provides functionality for confusion_stats operations.
}
\examples{
  cm <- matrix(c(10, 2, 1, 15), nrow=2)
  try({ confusion_stats(cm) })

}
\seealso{
\code{\link{VectorForgeML-package}}
}
