% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{drop_constant_columns}
\alias{drop_constant_columns}
\title{Drop Constant Columns}
\usage{
drop_constant_columns(X, eps = 1e-12)
}
\arguments{
\item{X}{input matrix/dataframe}

\item{eps}{for param eps}
}
\value{
cleaned matrix
}
\description{
Removes columns with zero variance.
}
\details{
Provides functionality for drop_constant_columns operations.
}
\examples{
  x <- data.frame(a=c(1,1,1), b=c(1,2,3))
  drop_constant_columns(x)

}
\seealso{
\code{\link{VectorForgeML-package}}
}
