% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoders.R
\docType{class}
\name{LabelEncoder-class}
\alias{LabelEncoder-class}
\alias{LabelEncoder}
\title{Label Encoder}
\value{
LabelEncoder object
}
\description{
Converts categorical labels into numeric values.
}
\details{
Provides functionality for LabelEncoder operations.
}

\examples{
  enc <- LabelEncoder$new()
  x <- c("a", "b", "a")
  enc$fit(x)
  enc$transform(x)

}
\seealso{
\code{\link{VectorForgeML-package}}
}
