% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\docType{class}
\name{LinearRegression-class}
\alias{LinearRegression-class}
\alias{LinearRegression}
\title{Linear Regression Model}
\value{
LinearRegression object
}
\description{
Fast linear regression implemented in C++ backend.
}
\details{
Provides functionality for LinearRegression operations.
}

\examples{
model <- LinearRegression$new()
X <- matrix(rnorm(100),50,2)
y <- rnorm(50)
model$fit(X,y)
model$predict(X)


}
\seealso{
\code{\link{VectorForgeML-package}}
}
