% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{macro_recall}
\alias{macro_recall}
\title{Macro Precision}
\usage{
macro_recall(y_true, y_pred)
}
\arguments{
\item{y_true}{true labels}

\item{y_pred}{predicted labels}
}
\value{
numeric score
}
\description{
Computes macro-averaged precision.
}
\details{
Provides functionality for macro_recall operations.
}
\seealso{
\code{\link{VectorForgeML-package}}
}
