% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mse}
\alias{mse}
\title{Mean Squared Error}
\usage{
mse(y_true, y_pred)
}
\arguments{
\item{y_true}{true values}

\item{y_pred}{predicted values}
}
\value{
numeric mse
}
\description{
Calculates regression error.
}
\details{
Provides functionality for mse operations.
}
\seealso{
\code{\link{VectorForgeML-package}}
}
