% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoders.R
\docType{class}
\name{OneHotEncoder-class}
\alias{OneHotEncoder-class}
\alias{OneHotEncoder}
\title{One Hot Encoder}
\value{
OneHotEncoder object
}
\description{
Converts categorical variables into binary vectors.
}
\details{
Provides functionality for OneHotEncoder operations.
}

\examples{
  enc <- OneHotEncoder$new()
  df <- data.frame(a=c("x","y","x"))
  enc$fit(df)
  enc$transform(df)

}
\seealso{
\code{\link{VectorForgeML-package}}
}
