% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\docType{class}
\name{Pipeline-class}
\alias{Pipeline-class}
\alias{Pipeline}
\title{Pipeline}
\value{
Pipeline object
}
\description{
Chains preprocessing and model steps.
}
\details{
Provides functionality for Pipeline operations.
}

\examples{
  model <- Pipeline$new(list(StandardScaler$new()))

}
\seealso{
\code{\link{VectorForgeML-package}}
}
