% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{Plot Confusion Matrix}
\usage{
plot_confusion_matrix(cm, normalize = TRUE)
}
\arguments{
\item{cm}{confusion matrix}

\item{normalize}{Normlize}
}
\value{
plot
}
\description{
Visualizes confusion matrix.
}
\details{
Provides functionality for plot_confusion_matrix operations.
}
\examples{
  cm <- matrix(c(10, 2, 1, 15), nrow=2)
  try({ plot_confusion_matrix(cm) })

}
\seealso{
\code{\link{VectorForgeML-package}}
}
