% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rmse}
\alias{rmse}
\title{Root Mean Squared Error}
\usage{
rmse(y_true, y_pred)
}
\arguments{
\item{y_true}{true values}

\item{y_pred}{predicted values}
}
\value{
numeric rmse
}
\description{
Square root of MSE.
}
\details{
Provides functionality for rmse operations.
}
\seealso{
\code{\link{VectorForgeML-package}}
}
