% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{train_test_split}
\alias{train_test_split}
\title{Train Test Split}
\usage{
train_test_split(X, y, test_size = 0.2, seed = NULL)
}
\arguments{
\item{X}{features}

\item{y}{labels}

\item{test_size}{proportion for test set}

\item{seed}{for random seed}
}
\value{
list
}
\description{
Splits dataset into training and testing sets.
}
\details{
Provides functionality for train_test_split operations.
}
\examples{
  X <- matrix(rnorm(20), nrow=10)
  y <- sample(0:1, 10, replace=TRUE)
  train_test_split(X, y, test_size=0.2)

}
\seealso{
\code{\link{VectorForgeML-package}}
}
